Allow disabling parent directory creation
By setting the `create_directories` pillar key in a certificate to
`False`, the acme state will not create a `file.directory` state for the
directories where the private key and certificate are placed. This might
be needed if another states creates this directory, to avoid a duplicate
state ID error.

Note: This option will *not* remove the requirements from the private
key and certificate states on the parent directory. You must define your
own file.directory state with the name of the parent directories.